Separate Django models from Event classes.
